More sane, but still crazy.
